move rename wallet to wallet settings
