+++ WalletWasabi.Fluent/Icons/Icons.axaml
@@ -513,8 +513,6 @@
       <StreamGeometry x:Key="undo_icon_regular">M35.795 9.83982C31.6214 5.23784 24.4923 5.33466 19.94 10.5101L19.923 10.5294L11.3645 19.4882H23.7441C24.4344 19.4882 24.9941 20.0478 24.9941 20.7382C24.9941 21.4285 24.4344 21.9882 23.7441 21.9882H8.19922C7.50887 21.9882 6.94922 21.4285 6.94922 20.7382V5.26886C6.94922 4.5785 7.50887 4.01886 8.19922 4.01886C8.88958 4.01886 9.44922 4.5785 9.44922 5.26886V17.8739L18.0802 8.83922C23.5294 2.66488 32.387 2.36064 37.6468 8.16035C42.8867 13.9381 41.2689 21.4757 37.6329 25.3549L37.6237 25.3648L28.6447 34.5376L28.6396 34.5428L19.8896 43.3796L19.8853 43.384L19.6343 43.6349C19.1462 44.1231 18.3547 44.1231 17.8666 43.6349C17.3784 43.1468 17.3784 42.3553 17.8666 41.8672L18.1132 41.6206L18.1175 41.6162L26.8581 32.7889L26.8632 32.7837L35.8178 23.6357C38.7365 20.5121 39.9898 14.4652 35.795 9.83982Z</StreamGeometry>
       <StreamGeometry x:Key="action_center_regular">F1 M14 21C15.1046 21 16 21.8954 16 23C16 24.1046 15.1046 25 14 25C12.8954 25 12 24.1046 12 23C12 21.8954 12.8954 21 14 21ZM23 21C24.1046 21 25 21.8954 25 23C25 24.1046 24.1046 25 23 25C21.8954 25 21 24.1046 21 23C21 21.8954 21.8954 21 23 21ZM5 21C6.10457 21 7 21.8954 7 23C7 24.1046 6.10457 25 5 25C3.89543 25 3 24.1046 3 23C3 21.8954 3.89543 21 5 21ZM14 12C15.1046 12 16 12.8954 16 14C16 15.1046 15.1046 16 14 16C12.8954 16 12 15.1046 12 14C12 12.8954 12.8954 12 14 12ZM23 12C24.1046 12 25 12.8954 25 14C25 15.1046 24.1046 16 23 16C21.8954 16 21 15.1046 21 14C21 12.8954 21.8954 12 23 12ZM5 12C6.10457 12 7 12.8954 7 14C7 15.1046 6.10457 16 5 16C3.89543 16 3 15.1046 3 14C3 12.8954 3.89543 12 5 12ZM14 3C15.1046 3 16 3.89543 16 5C16 6.10457 15.1046 7 14 7C12.8954 7 12 6.10457 12 5C12 3.89543 12.8954 3 14 3ZM23 3C24.1046 3 25 3.89543 25 5C25 6.10457 24.1046 7 23 7C21.8954 7 21 6.10457 21 5C21 3.89543 21.8954 3 23 3ZM5 3C6.10457 3 7 3.89543 7 5C7 6.10457 6.10457 7 5 7C3.89543 7 3 6.10457 3 5C3 3.89543 3.89543 3 5 3Z</StreamGeometry>
 
-      <StreamGeometry x:Key="rename_regular_20">M17.1813 2.92621C16.0291 1.71436 14.1047 1.69009 12.9222 2.87248L3.54741 12.2469C3.21958 12.5747 2.99204 12.9892 2.89148 13.4418L2.01387 17.3917C1.97678 17.5586 2.02754 17.7328 2.14844 17.8537C2.26934 17.9746 2.44362 18.0253 2.6105 17.9882L6.53689 17.115C7.00432 17.0111 7.43243 16.776 7.77103 16.4374L17.129 7.07934C18.27 5.93831 18.2933 4.09563 17.1813 2.92621ZM13.6293 3.5796C14.4143 2.7947 15.6917 2.81081 16.4566 3.61527C17.1948 4.39156 17.1793 5.61479 16.4219 6.37224L15.7507 7.0435L12.958 4.25087L13.6293 3.5796ZM12.2509 4.95796L15.0436 7.75061L7.06391 15.7303C6.85976 15.9345 6.60164 16.0762 6.31982 16.1389L3.1605 16.8414L3.86768 13.6587C3.92698 13.3918 4.06117 13.1473 4.2545 12.954L12.2509 4.95796Z</StreamGeometry>
-
       <DrawingImage x:Key="tagcontrol_dismiss">
         <DrawingImage.Drawing>
           <DrawingGroup>
