+++ WalletWasabi.Fluent/ViewModels/Wallets/WalletSettingsViewModel.cs
@@ -1,8 +1,12 @@
+using System.Reactive;
 using System.Reactive.Linq;
 using System.Windows.Input;
 using ReactiveUI;
 using WalletWasabi.Blockchain.Keys;
+using WalletWasabi.Fluent.Helpers;
+using WalletWasabi.Fluent.Validation;
 using WalletWasabi.Fluent.ViewModels.Navigation;
+using WalletWasabi.Models;
 using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.ViewModels.Wallets;
@@ -12,6 +16,7 @@ public partial class WalletSettingsViewModel : RoutableViewModel
 	private readonly Wallet _wallet;
 	[AutoNotify] private string _plebStopThreshold;
 	[AutoNotify] private bool _preferPsbtWorkflow;
+	[AutoNotify] private string _walletName;
 
 	public WalletSettingsViewModel(WalletViewModelBase walletViewModelBase)
 	{
@@ -25,7 +30,11 @@ public partial class WalletSettingsViewModel : RoutableViewModel
 
 		SetupCancel(enableCancel: false, enableCancelOnEscape: true, enableCancelOnPressed: true);
 
-		NextCommand = CancelCommand;
+		var canExecute =
+			this.WhenAnyValue(x => x.WalletName)
+				.Select(x => !string.IsNullOrWhiteSpace(x) && !Validations.Any);
+
+		NextCommand = ReactiveCommand.Create(() => OnNext(walletViewModelBase), canExecute);
 
 		VerifyRecoveryWordsCommand =
 			ReactiveCommand.Create(() => Navigate().To(new VerifyRecoveryWordsViewModel(_wallet)));
@@ -39,6 +48,9 @@ public partial class WalletSettingsViewModel : RoutableViewModel
 					_wallet.KeyManager.ToFile();
 					walletViewModelBase.RaisePropertyChanged(nameof(walletViewModelBase.PreferPsbtWorkflow));
 				});
+
+		_walletName = walletViewModelBase.WalletName;
+        this.ValidateProperty(x => x.WalletName, ValidateWalletName);
 	}
 
 	public bool IsHardwareWallet { get; }
@@ -48,4 +60,21 @@ public partial class WalletSettingsViewModel : RoutableViewModel
 	public override sealed string Title { get; protected set; }
 
 	public ICommand VerifyRecoveryWordsCommand { get; }
+
+	private void OnNext(WalletViewModelBase walletViewModelBase)
+	{
+		walletViewModelBase.WalletName = WalletName;
+		Title = $"{walletViewModelBase.WalletName} - Wallet Settings";
+
+		CancelCommand.Execute(Unit.Default);
+	}
+
+	private void ValidateWalletName(IValidationErrors errors)
+	{
+		var error = WalletHelpers.ValidateWalletName(WalletName);
+		if (error is { } e)
+		{
+			errors.Add(e.Severity, e.Message);
+		}
+	}
 }
