--- WalletWasabi.Fluent/Views/Wallets/WalletRenameView.axaml.cs
@@ -1,17 +1,0 @@
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace WalletWasabi.Fluent.Views.Wallets;
-
-public class WalletRenameView : UserControl
-{
-	public WalletRenameView()
-	{
-		InitializeComponent();
-	}
-
-	private void InitializeComponent()
-	{
-		AvaloniaXamlLoader.Load(this);
-	}
-}
