+++ WalletWasabi.Fluent/Views/Wallets/WalletSettingsView.axaml
@@ -5,6 +5,7 @@
              xmlns:c="using:WalletWasabi.Fluent.Controls"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
+             xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors"
              x:DataType="wallets:WalletSettingsViewModel"
              x:CompileBindings="True"
              x:Class="WalletWasabi.Fluent.Views.Wallets.WalletSettingsView">
@@ -14,6 +15,19 @@
                  EnableBack="{Binding EnableBack}">
 
     <StackPanel Spacing="20" Margin="0 30">
+
+
+      <StackPanel Classes="settingsLayout">
+        <DockPanel>
+          <TextBlock Text="Rename Wallet" />
+          <TextBox MaxLength="50" Text="{Binding WalletName, Mode=TwoWay}" Watermark="Type in a wallet name">
+            <Interaction.Behaviors>
+              <behaviors:FocusOnAttachedBehavior />
+              <behaviors:TextBoxSelectAllTextBehavior />
+            </Interaction.Behaviors>
+          </TextBox>
+        </DockPanel>
+      </StackPanel>
 
       <StackPanel Classes="settingsLayout">
         <DockPanel IsVisible="{Binding IsHardwareWallet}">
@@ -26,7 +40,8 @@
         <TextBlock Classes="h6" Text="Have you checked your wallet backup?" />
         <c:InfoMessage>
           <DockPanel>
-            <Button Margin="50 0" Command="{Binding VerifyRecoveryWordsCommand}" Content="Verify Recovery Words" DockPanel.Dock="Right" />
+            <Button Margin="50 0" Command="{Binding VerifyRecoveryWordsCommand}" Content="Verify Recovery Words"
+                    DockPanel.Dock="Right" />
             <TextBlock Text="It is recommended that you verify your wallet Recovery Words from time to time." />
           </DockPanel>
         </c:InfoMessage>
