+++ WalletWasabi.Fluent/Views/Wallets/WalletView.axaml
@@ -69,11 +69,6 @@
                       <PathIcon Data="{StaticResource stats_wallet_regular}" />
                     </MenuItem.Icon>
                   </MenuItem>
-                  <MenuItem Header="Rename Wallet" Command="{Binding RenameCommand}">
-                    <MenuItem.Icon>
-                      <PathIcon Data="{StaticResource rename_regular_20}" />
-                    </MenuItem.Icon>
-                  </MenuItem>
                 </MenuFlyout>
               </Button.Flyout>
               <PathIcon Data="{StaticResource more_regular}" Foreground="{DynamicResource TextForegroundColor}" />
